backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.13"
dependencies = [
    "django>=5.2.7",
    "django-cors-headers==4.9.0",
    "djangorestframework==3.16.1",
    "pillow==12.0.0",
    "psycopg2-binary==2.9.11",
    "python-decouple==3.8",
]
